backend/requirements.txt
fastapi[standard]>=0.122.0
geopy>=2.4.1
jupyter>=1.1.1
lightgbm>=4.6.0
matplotlib>=3.10.7
numpy>=2.3.5
pandas>=2.3.3
plotly>=6.5.0
prophet>=1.2.1
psycopg2-binary>=2.9.11
pydantic>=2.12.4
pytest>=9.0.1
python-dotenv>=1.2.1
pyyaml>=6.0.3
requests>=2.32.5
ruff>=0.14.7
scikit-learn>=1.7.2
seaborn>=0.13.2
sqlalchemy>=2.0.44
statsmodels>=0.14.5
streamlit>=1.51.0
supabase>=2.24.0
tqdm>=4.67.1
uvicorn[standard]>=0.38.0
xgboost>=3.1.2
